src/services/inventory-sync/requirements.txt
locust
websocket-client

